+++ eliot/prettyprint.py
@@ -34,6 +34,10 @@ if PY2:
     pprint.repr = _nicer_unicode_repr
 
 
+# Fields that all Eliot messages are expected to have:
+REQUIRED_FIELDS = {TASK_LEVEL_FIELD, TASK_UUID_FIELD, TIMESTAMP_FIELD}
+
+
 def pretty_format(message):
     """
     Convert a message dictionary into a human-readable string.
@@ -95,7 +99,15 @@ def _main():
         stdout.write(_CLI_HELP)
         raise SystemExit()
     for line in stdin:
-        message = loads(line)
+        try:
+            message = loads(line)
+        except ValueError:
+            stdout.write("Not JSON: {}\n\n".format(line.rstrip(b"\n")))
+            continue
+        if REQUIRED_FIELDS - set(message.keys()):
+            stdout.write("Not an Eliot message: {}\n\n".format(
+                line.rstrip(b"\n")))
+            continue
         result = pretty_format(message) + "\n"
         if PY2:
             result = result.encode("utf-8")