+++ docs/source/news.rst
@@ -9,6 +9,9 @@ Features:
 * Eliot tasks can now more easily :ref:`span multiple threads <cross thread tasks>` using the new ``eliot.preserve_context`` API.
 * ``eliot-prettyprint`` command line tool now pretty prints field values in a more informative manner.
 
+Bug fixes:
+
+* ``eliot-prettyprint`` now handles unparseable lines by skipping formatting them rather than exiting.
 
 0.10.1
 ^^^^^^